requirements.txt
Django==5.2.3
djangorestframework==3.16.0
gradio>=4.38.0